+++ manager/controllers/step_controller.go
@@ -156,8 +156,8 @@ func (r *StepReconciler) Reconcile(ctx context.Context, req ctrl.Request) (ctrl.
 		annotations[dfv1.KeyReplica] = strconv.Itoa(replica)
 		annotations[dfv1.KeyHash] = hash
 		annotations[dfv1.KeyDefaultContainer] = dfv1.CtrMain
-		annotations[dfv1.KeyKillCmd(dfv1.CtrMain)] = util.MustJSON([]string{dfv1.PathKill, "1"})
-		annotations[dfv1.KeyKillCmd(dfv1.CtrSidecar)] = util.MustJSON([]string{dfv1.PathKill, "1"})
+		annotations[dfv1.KeyKillCmd(dfv1.CtrMain)] = util.MustJSON([]string{dfv1.PathVarRun, "1"})
+		annotations[dfv1.KeyKillCmd(dfv1.CtrSidecar)] = util.MustJSON([]string{dfv1.PathVarRun, "1"})
 		if err := r.Client.Create(
 			ctx,
 			&corev1.Pod{