+++ runner/sidecar/handler.go
@@ -1,6 +1,8 @@
 package sidecar
 
 import (
+	"sync"
+
 	"github.com/Shopify/sarama"
 )
 
@@ -8,10 +10,24 @@ type handler struct {
 	f         func([]byte) error
 	partition int32
 	offset    int64
+	wg        sync.WaitGroup
 }
 
-func (*handler) Setup(_ sarama.ConsumerGroupSession) error   { return nil }
-func (*handler) Cleanup(_ sarama.ConsumerGroupSession) error { return nil }
+func (h *handler) Setup(_ sarama.ConsumerGroupSession) error {
+	h.wg.Add(1)
+	return nil
+}
+
+func (h *handler) Cleanup(_ sarama.ConsumerGroupSession) error {
+	h.wg.Done()
+	return nil
+}
+
+func (h *handler) Close() error {
+	h.wg.Wait()
+	return nil
+}
+
 func (h *handler) ConsumeClaim(sess sarama.ConsumerGroupSession, claim sarama.ConsumerGroupClaim) error {
 	for m := range claim.Messages() {
 		h.partition = m.Partition